Plan: plans/PR-Content-Ops-Brand-Voice-Editor-Extraction.md Slice phase: Product polish
This closes the settings-page deferred item by extracting the already-shipped
New Run brand voice selector/editor into a shared `BrandVoiceProfileManager`
component, then rendering that same component on `/content-ops/brand-voice`.
Settings can now create, edit, archive, refresh, and sample-import saved brand
voice profiles without sending operators back to New Run, while New Run keeps
using the same component for run profile selection.

## Intentional
- This does not add URL deep-linking from settings cards into New Run; the
  management detour is removed instead.
- Settings keeps scan-friendly inventory cards and uses one shared manager
  above them rather than putting duplicate mutation controls on every card.
- The diff exceeds the 400 LOC soft cap because this is a move-heavy extraction
  of the existing editor, not a second implementation.

## Deferred
- None.

## Parked hardening
- None.

## Verification
- `cd atlas-intel-ui && npm ci` - installed locked UI dependencies in the new
  worktree; existing npm audit findings reported, no dependency changes made.
- `cd atlas-intel-ui && npm run test:content-ops-brand-voice-profile-selector`
  - 12/12 tests passed.
- `cd atlas-intel-ui && npm run test:content-ops-brand-voice-settings-page`
  - 4/4 tests passed.
- `cd atlas-intel-ui && npm run lint` - passed.
- `cd atlas-intel-ui && npm run build` - passed.
- Manual workflow grep for `test:content-ops-brand-voice-profile-selector` and
  `test:content-ops-brand-voice-settings-page` in
  `.github/workflows/atlas_intel_ui_checks.yml` - both enrolled.
- `bash scripts/local_pr_review.sh --current-pr-body-file tmp/pr-body-content-ops-brand-voice-editor-extraction.md`
  - passed.

## Diff size
6 files, +872 / -691.
