+++ atlas-intel-ui/scripts/content-ops-brand-voice-settings-page.test.mjs
@@ -11,6 +11,10 @@ const pageSource = readFileSync(
   new URL('../src/pages/ContentOpsBrandVoiceSettings.tsx', import.meta.url),
   'utf8',
 )
+const managerSource = readFileSync(
+  new URL('../src/components/contentOps/BrandVoiceProfileManager.tsx', import.meta.url),
+  'utf8',
+)
 
 test('brand voice settings page is registered as protected Content Ops route', () => {
   assert.ok(
@@ -37,16 +41,22 @@ test('settings page lists tenant brand voice profiles with real API and states',
   assert.ok(pageSource.includes('useApiData(() => fetchContentOpsBrandVoiceProfiles(), [])'))
   assert.ok(pageSource.includes('if ((loading || refreshing) && !profiles)'))
   assert.ok(pageSource.includes('Loading brand voice profiles...'))
+  assert.ok(pageSource.includes('BrandVoiceProfileManager'))
+  assert.ok(pageSource.includes('selectedProfileId'))
+  assert.ok(pageSource.includes('onChange={setSelectedProfileId}'))
   assert.ok(pageSource.includes('No saved brand voice profiles'))
   assert.ok(pageSource.includes('Refresh failed:'))
   assert.ok(pageSource.includes('profile.descriptors'))
   assert.ok(pageSource.includes('profile.exemplars[0]'))
 })
 
-test('settings page sends create/edit/run actions back to New Run editor', () => {
+test('settings page uses shared manager instead of routing management to New Run', () => {
   const newRunLinks = pageSource.match(/to="\/content-ops\/new"/g) ?? []
-  assert.ok(newRunLinks.length >= 2)
-  assert.ok(pageSource.includes('New profile'))
-  assert.ok(pageSource.includes('Open New Run'))
+  assert.equal(newRunLinks.length, 0)
+  assert.ok(pageSource.includes("selected ? 'Selected' : 'Manage'"))
+  assert.ok(managerSource.includes('New brand voice'))
+  assert.ok(managerSource.includes('Edit brand voice'))
+  assert.ok(managerSource.includes('Archive'))
+  assert.ok(managerSource.includes('fetchContentOpsBrandVoiceSampleUrl'))
   assert.equal(pageSource.includes('Use in run'), false)
 })
